backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Image Processing
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3

# OCR
pytesseract==0.3.10

# OpenAI
openai==1.10.0
tiktoken==0.5.2

# Storage
boto3==1.34.25
python-multipart==0.0.6

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utils
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0
